requirements.txt
# Core ML/Embeddings
FlagEmbedding>=1.2.0
torch>=2.0.0
numpy>=1.24.0

# Vector Search
faiss-cpu>=1.7.4

# Data Validation
pydantic>=2.0.0

# Web Scraping & APIs
feedparser>=6.0.0
requests>=2.31.0

# Gemini API
google-genai>=0.2.0

# Environment Variables
python-dotenv>=1.0.0
